requirements.txt
# ============================================================
# Webtoon RAG Recommendation System - Final Requirements
# ============================================================

# -----------------------------
# Core ML, Embeddings & NLP
# -----------------------------
sentence-transformers==3.3.1
torch>=2.2.0
numpy>=1.24.0,<2.0.0
transformers>=4.30.0

# -----------------------------
# Database & Vector Store
# -----------------------------
supabase>=2.3.0
vecs>=0.4.0

# -----------------------------
# LLM Integration
# -----------------------------
google-generativeai>=0.8.0

# -----------------------------
# Web Framework (if you run API/UI)
# -----------------------------
Flask==3.0.0
Flask-CORS==4.0.0
Werkzeug>=3.0.0

# -----------------------------
# Config & Validation
# -----------------------------
python-dotenv>=1.0.0
pydantic>=2.0.0

# -----------------------------
# Data Processing
# -----------------------------
pandas>=2.0.0

# -----------------------------
# Utilities
# -----------------------------
requests>=2.31.0
rich>=13.0.0

# -----------------------------
# Optional: Development Tools
# -----------------------------
# black>=23.0.0
# pytest>=7.4.0
# pytest-asyncio>=0.23.0
# pytest-cov>=4.1.0
